+++ tests/checks/layout_test.py
@@ -1,3 +1,4 @@
+from fontTools.ttLib import TTFont
 from fontbakery.status import FAIL
 from fontbakery.codetesting import (
     assert_PASS,
@@ -11,9 +12,17 @@ def test_check_layout_valid_feature_tags():
     """Does the font have any invalid feature tags?"""
     check = CheckTester("com.google.fonts/check/layout_valid_feature_tags")
 
+    # test font with valid, registered feature tags.
     font = TEST_FILE("nunito/Nunito-Regular.ttf")
     assert_PASS(check(font))
 
+    # test font with valid, private use feature tags.
+    # change font's feature tag to have non-registered, all uppercase private tags
+    font_obj = TTFont(font)
+    font_obj["GSUB"].table.FeatureList.FeatureRecord[0].FeatureTag = "TEST"
+    assert_PASS(check(font_obj))
+
+    # test font with invalid feature tags: not registered, and not all uppercase.
     font = TEST_FILE("rosarivo/Rosarivo-Regular.ttf")
     assert_results_contain(check(font), FAIL, "bad-feature-tags")