+++ CHANGELOG.md
@@ -8,6 +8,10 @@ A more detailed list of changes is available in the corresponding milestones for
     - v0.12.0a1 (2024-Feb-14)
     - v0.12.0a2 (2024-Feb-21)
 
+### Changes to existing checks
+#### On the Open Type Profile
+  - **[com.google.fonts/check/layout_valid_feature_tags]:** Updated the check to allow valid private-use feature tags. (issue #4544)
+
 
 ## 0.12.0a2 (2024-Feb-21)
   - **[multi-threading]:** Lock reporting when returning results. (issue #4494)